shattered btutils into  individual function files
