+++ src/utils/constants.js
@@ -1,0 +1,304 @@
+export const btnList = [
+    /*Default                      Keyboard             Mouse                PS3                  PS4 / PS5            Wiimote              Wiimote + Classic    Wiimote + Nunchuck   WiiU / Switch Pro    Switch NES           Switch SNES          Switch MD / Genesis  Switch N64           Switch Joycon        Xbox One S / X|S     Steam                NeoGeo (Parallel 1P) PCE                  PCE 6 btns           NES                  SNES                 CD-i                 JVS                  3DO                  Jaguar         Jaguar 6D             PC-FX                VB                          N64                         GameCube             Atari / SMS          MD / Genesis         Saturn               Dreamcast            PSX / PS2           */
+    ['GP: LX Left;  KB: A',       'KB: A',             '',                  'Left stick left',   'Left stick left',   '',                '[C] Left stick left',  '[N] Stick left',   'Left stick left',   '',                  '',                  '',                  'Stick left',        'Stick left (H)',    'Left stick left',   'Left pad left',     '',                  '',                  '',                  '',                  '',                  'Stick left',        'Stick left',        '*Stick left',       '8 (Y)',   'Left stick left (X)',    '',                  '',                         'Stick left',               'Left stick left',   '',                  '',                  '*Stick left',       'Left stick left',   'Left stick left',  ],
+    ['GP: LX Right; KB: D',       'KB: D',             '',                  'Left stick right',  'Left stick right',  '',                '[C] Left stick right', '[N] Stick right',  'Left stick right',  '',                  '',                  '',                  'Stick right',       'Stick right (H)',   'Left stick right',  'Left pad right',    '',                  '',                  '',                  '',                  '',                  'Stick right',       'Stick right',       '*Stick right',      '6',       'Left stick right (X)',   '',                  '',                         'Stick right',              'Left stick right',  '',                  '',                  '*Stick right',      'Left stick right',  'Left stick right', ],
+    ['GP: LY Down;  KB: S',       'KB: S',             '',                  'Left stick down',   'Left stick down',   '',                '[C] Left stick down',  '[N] Stick down',   'Left stick down',   '',                  '',                  '',                  'Stick down',        'Stick down (H)',    'Left stick down',   'Left pad down',     '',                  '',                  '',                  '',                  '',                  'Stick down',        'Stick down',        '*Stick down',       '7 (Z)',   'Left stick down (Y)',    '',                  '',                         'Stick down',               'Left stick down',   '',                  '',                  '*Stick down',       'Left stick down',   'Left stick down',  ],
+    ['GP: LY Up;    KB: W',       'KB: W',             '',                  'Left stick up',     'Left stick up',     '',                '[C] Left stick up',    '[N] Stick up',     'Left stick up',     '',                  '',                  '',                  'Stick up',          'Stick up (H)',      'Left stick up',     'Left pad up',       '',                  '',                  '',                  '',                  '',                  'Stick up',          'Stick up',          '*Stick up',         '5',       'Left stick up (Y)',      '',                  '',                         'Stick up',                 'Left stick up',     '',                  '',                  '*Stick up',         'Left stick up',     'Left stick up',    ],
+    ['GP: RX Left;  M: X Left',   '',                  'M: X Left',         'Right stick left',  'Right stick left',  '',                '[C] Right stick left', '',                 'Right stick left',  '',                  '',                  '',                  'C-Left',            '',                  'Right stick left',  'Right pad left',    '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '4',       'Right stick left (TX)',  '',                  'Right D-pad left',         'C-Left',                   'Right stick left',  '',                  '',                  '',                  '*Right stick left', 'Right stick left', ],
+    ['GP: RX Right; M: X Right',  '',                  'M: X Right',        'Right stick right', 'Right stick right', '',                '[C] Right stick right','',                 'Right stick right', '',                  '',                  '',                  'C-Right',           '',                  'Right stick right', 'Right pad right',   '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '2',       'Right stick right (TX)', '',                  'Right D-pad right',        'C-Right',                  'Right stick right', '',                  '',                  '',                  '*Right stick right','Right stick right',],
+    ['GP: RY Down;  M: Y Down',   '',                  'M: Y Down',         'Right stick down',  'Right stick down',  '',                '[C] Right stick down', '',                 'Right stick down',  '',                  '',                  '',                  'C-Down',            '',                  'Right stick down',  'Right pad down',    '',                  '',                  '',                  '',                  '',                  '',                  '',                  '*Throttle down',    '3',       'Right stick down (TY)',  '',                  'Right D-pad down',         'C-Down',                   'Right stick down',  '',                  '',                  '',                  '*Right stick down', 'Right stick down', ],
+    ['GP: RY Up;    M: Y Up',     '',                  'M: Y Up',           'Right stick up',    'Right stick up',    '',                '[C] Right stick up',   '',                 'Right stick up',    '',                  '',                  '',                  'C-Up',              '',                  'Right stick up',    'Right pad up',      '',                  '',                  '',                  '',                  '',                  '',                  '',                  '*Throttle up',      '1',       'Right stick up (TY)',    '',                  'Right D-pad up',           'C-Up',                     'Right stick up',    '',                  '',                  '',                  '*Right stick up',   'Right stick up',   ],
+    ['GP: LD Left;  KB: Left',    'KB: Left',          '',                  'D-pad left',        'D-pad left',        'D-pad left (H)',  '[C] D-pad left',     '[W] D-pad left (V)', 'D-pad left',        'D-pad left',        'D-pad left',        'D-pad left',        'D-pad left',        '',                  'D-pad left',        'Stick left',        'D-pad left',        'D-pad left',        'D-pad left',        'D-pad left',        'D-pad left',        '',                  'D-pad left',        'D-pad left',        'D-pad left',  'D-pad left',         'D-pad left',        'Left D-pad left',          'D-pad left',               'D-pad left',        'D-pad left',        'D-pad left',        'D-pad left',        'Left D-pad left',   'D-pad left',       ],
+    ['GP: LD Right; KB: Right',   'KB: Right',         '',                  'D-pad right',       'D-pad right',       'D-pad right (H)', '[C] D-pad right',    '[W] D-pad right (V)','D-pad right',       'D-pad right',       'D-pad right',       'D-pad right',       'D-pad right',       '',                  'D-pad right',       'Stick right',       'D-pad right',       'D-pad right',       'D-pad right',       'D-pad right',       'D-pad right',       '',                  'D-pad right',       'D-pad right',       'D-pad right', 'D-pad right',        'D-pad right',       'Left D-pad right',         'D-pad right',              'D-pad right',       'D-pad right',       'D-pad right',       'D-pad right',       'Left D-pad right',  'D-pad right',      ],
+    ['GP: LD Down;  KB: Down',    'KB: Down',          '',                  'D-pad down',        'D-pad down',        'D-pad down (H)',  '[C] D-pad down',     '[W] D-pad down (V)', 'D-pad down',        'D-pad down',        'D-pad down',        'D-pad down',        'D-pad down',        '',                  'D-pad down',        'Stick down',        'D-pad down',        'D-pad down',        'D-pad down',        'D-pad down',        'D-pad down',        '',                  'D-pad down',        'D-pad down',        'D-pad down',  'D-pad down',         'D-pad down',        'Left D-pad down',          'D-pad down',               'D-pad down',        'D-pad down',        'D-pad down',        'D-pad down',        'Left D-pad down',   'D-pad down',       ],
+    ['GP: LD Up;    KB: Up',      'KB: Up',            '',                  'D-pad up',          'D-pad up',          'D-pad up (H)',    '[C] D-pad up',       '[W] D-pad up (V)',   'D-pad up',          'D-pad up',          'D-pad up',          'D-pad up',          'D-pad up',          '',                  'D-pad up',          'Stick up',          'D-pad up',          'D-pad up',          'D-pad up',          'D-pad up',          'D-pad up',          '',                  'D-pad up',          'D-pad up',          'D-pad up',    'D-pad up',           'D-pad up',          'Left D-pad up',            'D-pad up',                 'D-pad up',          'D-pad up',          'D-pad up',          'D-pad up',          'Left D-pad up',     'D-pad up',         ],
+    ['GP: RD Left;  M: WX Left',  '',                  'M: WX Left',        '',                  'Mute',              '',                '[W] D-pad left (H)',  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  'L pad click',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '*',                  '',                  '',                         '',                         '',                  '',                  '',                  '',                  '*Right D-pad left', '',                 ],
+    ['GP: RD Right; M: WX Right', '',                  'M: WX Right',       '',                  '',                  '',                '[W] D-pad right (H)', '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  'R pad click',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '0',                  '',                  '',                         '',                         '',                  '',                  '',                  '',                  '*Right D-pad right','',                 ],
+    ['GP: RD Down;  M: WY Down',  '',                  'M: WY Down',        '',                  '',                  '',                '[W] D-pad down (H)',  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '#',                  '',                  '',                         '',                         '',                  '',                  '',                  '',                  '*Right D-pad down', '',                 ],
+    ['GP: RD Up;    M: WY Up',    '',                  'M: WY Up',          '',                  '',                  '',                '[W] D-pad up (H)',    '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '*Right D-pad up',   '',                 ],
+    ['GP: RB Left;  KB: Q; M: 4', 'KB: Q',             'M: 4',              'Square',            'Square',            '1',                 '[C] Y',             '[W] 1',             'Y',                 'B',                 'Y',                 'A',                 'B',                 'B | D-pad up',      'X',                 'X',                 'A',                 'II',                'III',               'B',                 'Y',                 '1',                 '3',                 'A',                 'C',           'C',                  'III',               'B',                        'B',                        'B',                 '1',                 'A',                 'A',                 'X',                 'Square',           ],
+    ['GP: RB Right; KB: R; M: 5', 'KB: R',             'M: 5',              'Circle',            'Circle',            'B',                 '[C] A',             '[W] 2',             'A',                 '',                  'A',                 'C',                 '',                  'X | D-pad down',    'B',                 'B',                 'D',                 '',                  'I',                 '',                  'A',                 '',                  '2',                 'C',                 'A',           'A',                  'I',                 '',                         'C-Down',                   'X',                 '',                  'C',                 'C',                 'B',                 'Circle',           ],
+    ['GP: RB Down;  KB: E; M: 6', 'KB: E',             'M: 6',              'X',                 'X',                 '2',                 '[C] B',             '[W] A',             'B',                 'A',                 'B',                 'B',                 'A',                 'A | D-pad left',    'A',                 'A',                 'B',                 'I',                 'II',                'A',                 'B',                 '2',                 '1',                 'B',                 'B',           'B',                  'II',                'A',                        'A',                        'A',                 '2',                 'B',                 'B',                 'A',                 'X',                ],
+    ['GP: RB Up;    KB: F; M: 7', 'KB: F',             'M: 7',              'Triangle',          'Triangle',          'A',                 '[C] X',             '',                  'X',                 '',                  'X',                 'Y',                 '',                  'Y | D-pad right',   'Y',                 'Y',                 'C',                 '',                  'V',                 '',                  'X',                 '',                  '4',                 '*Trigger',          '8 (Y)',       '8 (Y)',              'V',                 '',                         'C-Left',                   'Y',                 '',                  '*Y',                'Y',                 'Y',                 'Triangle',         ],
+    ['GP: MM; KB: Esc',           'KB: Esc',           '',                  'Start',             'OPTIONS',           '+',                 '[C] +',             '',                  '+',                 'Start',             'Start',             'Start',             'Start',             '+ | Capture',       'Menu',              'Start',             'Start',             'Run',               'Run',               'Start',             'Start',             '',                  'Start',             'P',                 'Option',      'Option',             'Run',               'Start',                    'Start',                    'Start',             '',                  'Start',             'Start',             'Start',             'Start',            ],
+    ['GP: MS; KB: Enter',         'KB: Enter',         '',                  'Select',            'SHARE',             '-',                 '[C] -',             '',                  '-',                 'Select',            'Select',            'Mode',              '',                  '- | Home',          'View',              'Back',              'Select',            'Select',            'Select',            'Select',            'Select',            '',                  'Coin',              'X',                 'Pause',       'Pause',              'Select',            'Select',                   '',                         '',                  '',                  '*Mode',             '',                  '*D',                'Select',           ],
+    ['GP: MT; KB: L Win',         'KB: L Win',         '',                  'PS',                'PS',                'Home',              '[C] Home',          '',                  'Home',              '',                  '',                  'Home',              'Home',              '',                  'Xbox',              'Steam',             'Extra',             '',                  '',                  '',                  '',                  '',                  'Service',           '',                  '0',           '2',                  '',                  'Start',                    '**Toggle Ctrl/Rumble pak', '',                  '',                  '',                  '',                  '',                  'Analog',           ],
+    ['GP: MQ; KB: Hash',          'KB: Hash',          '',                  '',                  'TouchPad',          '',                  '[W] B',             '',                  'Capture',           '',                  '',                  'Capture',           'Capture',           '',                  'Share',             'Stick click',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '5',                  '',                  'Select',                   '**Rotate memory bank',     '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['GP: LM; M: Right',          '',                  'M: Right',          'L2',                'L2',                '',                  '[C] Analog L',      '[N] Z',             'ZL',                'L',                 'L',                 '',                  'Z',                 'SL',                'LT',                'Analog LT',         '6',                 '',                  'IV',                '',                  'L',                 '',                  '5',                 'L',                 '4 (L)',       'Z Axis',             'IV',                'L',                        'Z',                        'Analog L',          '',                  '*X',                'L',                 'L',                 'L2',               ],
+    ['GP: LS; KB: Z',             'KB: Z',             '',                  'L1',                'L1',                '',                  '[C] ZL',            '[N] C',             'L',                 '',                  'ZL',                'X',                 'L',                 'L | R',             'LB',                'LB',                '6',                 '',                  'IV',                '',                  'L',                 '',                  '7',                 'L',                 '7 (Z)',       '4 (L)',              'IV',                'L',                        'L',                        'Z',                 '',                  '*X',                'X',                 '*Z',                'L1',               ],
+    ['GP: LT; KB: L CTRL',        'KB: L CTRL',        '',                  '',                  '',                  '',                  '[C] Digital L',     '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  'Digital LT',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '1',                  '',                  '',                         '',                         'Digital L',         '',                  '',                  '',                  '',                  '',                 ],
+    ['GP: LJ; M: Middle',         '',                  'M: Middle',         'L3',                'L3',                '',                  '[W] 1',             '',                  'L Stick',           '',                  '',                  '',                  '',                  'Stick click',       'L Stick',           'L grip',            '',                  '',                  '',                  '',                  '',                  '',                  '9',                 '',                  '*',           '7 (Z)',              'Mode 1',            '**VTAP PALETTE',           '',                         '',                  '',                  '',                  '',                  '',                  'L3',               ],
+    ['GP: RM; M: Left',           '',                  'M: Left',           'R2',                'R2',                '',                  '[C] Analog R',      '[W] B',             'ZR',                'R',                 'R',                 '',                  'ZR',                'SR',                'RT',                'Analog RT',         'Credit',            '',                  'VI',                '',                  'R',                 '',                  '6',                 'R',                 '6 (R)',       'TZ Axis',            'VI',                'R',                        'Z',                        'Analog R',          '',                  '*Z',                'R',                 'R',                 'R2',               ],
+    ['GP: RS; KB: X; M: 8',       'KB: X',             '',                  'R1',                'R1',                '',                  '[C] ZR',            '',                  'R',                 '',                  'ZR',                'Z',                 'R',                 'ZL | ZR',           'RB',                'RB',                'Credit',            '',                  'VI',                '',                  'R',                 '',                  '8',                 'R',                 '9 (X)',       '6 (R)',              'VI',                'R',                        'R',                        'Z',                 '',                  '*Z',                'Z',                 '*C',                'R1',               ],
+    ['GP: RT; KB: L Shift',       'KB: L Shift',       '',                  '',                  '',                  '',                  '[C] Digital R',     '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  'Digital RT',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '3',                  '',                  '',                         '',                         'Digital R',         '',                  '',                  '',                  '',                  '',                 ],
+    ['GP: RJ; KB: Space',         'KB: Space',         '',                  'R3',                'R3',                '',                  '[W] 2',             '',                  'R Stick',           '',                  '',                  '',                  '',                  '',                  'R Stick',           'R grip',            '',                  '',                  '',                  '',                  '',                  '',                  '10',                '',                  '#',           '9 (X)',              'Mode 2',            '**VTAP MODE',              '',                         '',                  '',                  '',                  '',                  '',                  'R3',               ],
+    ['KB: B',                     'KB: B',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: C',                     'KB: C',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: G',                     'KB: G',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: H',                     'KB: H',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: I',                     'KB: I',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: J',                     'KB: J',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: K',                     'KB: K',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: L',                     'KB: L',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: M',                     'KB: M',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: N',                     'KB: N',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: O',                     'KB: O',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: P',                     'KB: P',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: T',                     'KB: T',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: U',                     'KB: U',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: V',                     'KB: V',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Y',                     'KB: Y',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 1',                     'KB: 1',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 2',                     'KB: 2',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 3',                     'KB: 3',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 4',                     'KB: 4',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 5',                     'KB: 5',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 6',                     'KB: 6',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 7',                     'KB: 7',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 8',                     'KB: 8',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 9',                     'KB: 9',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: 0',                     'KB: 0',             '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Backspace',             'KB: Backspace',     '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Tab',                   'KB: Tab',           '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Minus',                 'KB: Minus',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Equal',                 'KB: Equal',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: L Brace',               'KB: L Brace',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: R Brace',               'KB: R Brace',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Backslash',             'KB: Backslash',     '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Semicolon',             'KB: Semicolon',     '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Apostrophe',            'KB: Apostrophe',    '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Grave',                 'KB: Grave',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Comma',                 'KB: Comma',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Dot',                   'KB: Dot',           '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Slash',                 'KB: Slash',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Capslock',              'KB: Capslock',      '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F1',                    'KB: F1',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F2',                    'KB: F2',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F3',                    'KB: F3',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F4',                    'KB: F4',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F5',                    'KB: F5',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F6',                    'KB: F6',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F7',                    'KB: F7',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F8',                    'KB: F8',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F9',                    'KB: F9',            '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F10',                   'KB: F10',           '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F11',                   'KB: F11',           '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: F12',                   'KB: F12',           '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Print Screen',          'KB: Print Screen',  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Scroll',                'KB: Scroll',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Pause',                 'KB: Pause',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Insert',                'KB: Insert',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Home',                  'KB: Home',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Page Up',               'KB: Page Up',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Delete',                'KB: Delete',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: End',                   'KB: End',           '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Page Down',             'KB: Page Down',     '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: Numlock',               'KB: Numlock',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP Div',                'KB: KP Div',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP Multi',              'KB: KP Multi',      '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP Minus',              'KB: KP Minus',      '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP Plus',               'KB: KP Plus',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP Enter',              'KB: KP Enter',      '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 1',                  'KB: KP 1',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 2',                  'KB: KP 2',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 3',                  'KB: KP 3',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 4',                  'KB: KP 4',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 5',                  'KB: KP 5',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 6',                  'KB: KP 6',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 7',                  'KB: KP 7',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 8',                  'KB: KP 8',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 9',                  'KB: KP 9',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP 0',                  'KB: KP 0',          '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: KP Dot',                'KB: KP Dot',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: L Alt',                 'KB: L Alt',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: R Ctrl',                'KB: R Ctrl',        '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: R Shift',               'KB: R Shift',       '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: R Alt',                 'KB: R Alt',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+    ['KB: R Win',                 'KB: R Win',         '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',                  '',            '',                   '',                  '',                         '',                         '',                  '',                  '',                  '',                  '',                  '',                 ],
+];
+
+export const brUuid = [
+    "56830f56-5180-fab0-314b-2fa176799a00",
+    "56830f56-5180-fab0-314b-2fa176799a01",
+    "56830f56-5180-fab0-314b-2fa176799a02",
+    "56830f56-5180-fab0-314b-2fa176799a03",
+    "56830f56-5180-fab0-314b-2fa176799a04",
+    "56830f56-5180-fab0-314b-2fa176799a05",
+    "56830f56-5180-fab0-314b-2fa176799a06",
+    "56830f56-5180-fab0-314b-2fa176799a07",
+    "56830f56-5180-fab0-314b-2fa176799a08",
+    "56830f56-5180-fab0-314b-2fa176799a09",
+    "56830f56-5180-fab0-314b-2fa176799a0a",
+    "56830f56-5180-fab0-314b-2fa176799a0b",
+  ];
+  
+  export const block = 4096;
+  
+  export const pakSize = 32 * 1024;
+  
+  export const mtu = 244;
+  
+  export const ota_start = 0xa5;
+  
+  export const ota_abort = 0xde;
+  
+  export const ota_end = 0x5a;
+  
+  export const sys_deep_sleep = 0x37;
+  
+  export const sys_reset = 0x38;
+  
+  export const btn = {
+    PAD_LX_LEFT: 0,
+    PAD_LX_RIGHT: 1,
+    PAD_LY_DOWN: 2,
+    PAD_LY_UP: 3,
+    PAD_RX_LEFT: 4,
+    PAD_RX_RIGHT: 5,
+    PAD_RY_DOWN: 6,
+    PAD_RY_UP: 7,
+    PAD_LD_LEFT: 8,
+    PAD_LD_RIGHT: 9,
+    PAD_LD_DOWN: 10,
+    PAD_LD_UP: 11,
+    PAD_RD_LEFT: 12,
+    PAD_RD_RIGHT: 13,
+    PAD_RD_DOWN: 14,
+    PAD_RD_UP: 15,
+    PAD_RB_LEFT: 16,
+    PAD_RB_RIGHT: 17,
+    PAD_RB_DOWN: 18,
+    PAD_RB_UP: 19,
+    PAD_MM: 20,
+    PAD_MS: 21,
+    PAD_MT: 22,
+    PAD_MQ: 23,
+    PAD_LM: 24,
+    PAD_LS: 25,
+    PAD_LT: 26,
+    PAD_LJ: 27,
+    PAD_RM: 28,
+    PAD_RS: 29,
+    PAD_RT: 30,
+    PAD_RJ: 31,
+    MOUSE_X_LEFT: 4,
+    MOUSE_X_RIGHT: 5,
+    MOUSE_Y_DOWN: 6,
+    MOUSE_Y_UP: 7,
+    MOUSE_WX_LEFT: 12,
+    MOUSE_WX_RIGHT: 13,
+    MOUSE_WY_DOWN: 14,
+    MOUSE_WY_UP: 15,
+    MOUSE_LEFT: 28,
+    MOUSE_MIDDLE: 27,
+    MOUSE_RIGHT: 24,
+    MOUSE_4: 16,
+    MOUSE_5: 17,
+    MOUSE_6: 18,
+    MOUSE_7: 19,
+    MOUSE_8: 29,
+    KB_A: 0,
+    KB_D: 1,
+    KB_S: 2,
+    KB_W: 3,
+    KB_LEFT: 8,
+    KB_RIGHT: 9,
+    KB_DOWN: 10,
+    KB_UP: 11,
+    KB_Q: 16,
+    KB_R: 17,
+    KB_E: 18,
+    KB_F: 19,
+    KB_ESC: 20,
+    KB_ENTER: 21,
+    KB_LWIN: 22,
+    KB_HASH: 23,
+    KB_Z: 25,
+    KB_LCTRL: 26,
+    KB_X: 29,
+    KB_LSHIFT: 30,
+    KB_SPACE: 31,
+    KB_B: 32,
+    KB_C: 33,
+    KB_G: 34,
+    KB_H: 35,
+    KB_I: 36,
+    KB_J: 37,
+    KB_K: 38,
+    KB_L: 39,
+    KB_M: 40,
+    KB_N: 41,
+    KB_O: 42,
+    KB_P: 43,
+    KB_T: 44,
+    KB_U: 45,
+    KB_V: 46,
+    KB_Y: 47,
+    KB_1: 48,
+    KB_2: 49,
+    KB_3: 50,
+    KB_4: 51,
+    KB_5: 52,
+    KB_6: 53,
+    KB_7: 54,
+    KB_8: 55,
+    KB_9: 56,
+    KB_0: 57,
+    KB_BACKSPACE: 58,
+    KB_TAB: 59,
+    KB_MINUS: 60,
+    KB_EQUAL: 61,
+    KB_LEFTBRACE: 62,
+    KB_RIGHTBRACE: 63,
+    KB_BACKSLASH: 64,
+    KB_SEMICOLON: 65,
+    KB_APOSTROPHE: 66,
+    KB_GRAVE: 67,
+    KB_COMMA: 68,
+    KB_DOT: 69,
+    KB_SLASH: 70,
+    KB_CAPSLOCK: 71,
+    KB_F1: 72,
+    KB_F2: 73,
+    KB_F3: 74,
+    KB_F4: 75,
+    KB_F5: 76,
+    KB_F6: 77,
+    KB_F7: 78,
+    KB_F8: 79,
+    KB_F9: 80,
+    KB_F10: 81,
+    KB_F11: 82,
+    KB_F12: 83,
+    KB_PSCREEN: 84,
+    KB_SCROLL: 85,
+    KB_PAUSE: 86,
+    KB_INSERT: 87,
+    KB_HOME: 88,
+    KB_PAGEUP: 89,
+    KB_DEL: 90,
+    KB_END: 91,
+    KB_PAGE_DOWN: 92,
+    KB_NUMLOCK: 93,
+    KB_KP_DIV: 94,
+    KB_KP_MULTI: 95,
+    KB_KP_MINUS: 96,
+    KB_KP_PLUS: 97,
+    KB_KP_ENTER: 98,
+    KB_KP_1: 99,
+    KB_KP_2: 100,
+    KB_KP_3: 101,
+    KB_KP_4: 102,
+    KB_KP_5: 103,
+    KB_KP_6: 104,
+    KB_KP_7: 105,
+    KB_KP_8: 106,
+    KB_KP_9: 107,
+    KB_KP_0: 108,
+    KB_KP_DOT: 109,
+    KB_LALT: 110,
+    KB_RCTRL: 111,
+    KB_RSHIFT: 112,
+    KB_RALT: 113,
+    KB_RWIN: 114,
+  };
